+++ src/config/cli_args.rs
@@ -5,6 +5,7 @@ use log::error;
 use structopt::StructOpt;
 
 use crate::config;
+use crate::config::units::{Temperature, WindSpeed};
 use crate::location::CurrentLocation;
 use crate::location::ip_api::IpApi;
 use crate::location::LocationProvider;
@@ -16,29 +17,25 @@ use crate::weather::providers::WeatherProvider;
 #[structopt(author, about)]
 pub struct CliArgs {
     /// Enables verbose debug output
-    #[structopt(short, long)]
+    #[structopt(short = "d", long)]
     pub debug: bool,
-
+    /// Prints the current city
+    #[structopt(short = "C", long)]
+    pub current_city: bool,
     /// Prints the default config path
     #[structopt(short = "p", long)]
     pub default_config_path: bool,
     /// Path to an alternative config file
-    #[structopt(short, long)]
+    #[structopt(short = "c", long)]
     pub config_file: Option<String>,
 
     /// The format to display the weather status in
     ///
     /// Available tags:
     /// <icon>,
-    /// <temperature_kelvin>,
-    /// <kelvin_max>,
-    /// <kelvin_min>,
-    /// <temperature_celsius>,
-    /// <celsius_max>,
-    /// <celsius_min>,
-    /// <temperature_fahrenheit>,
-    /// <fahrenheit_max>,
-    /// <fahrenheit_min>,
+    /// <temperature>,
+    /// <temperature_max>,
+    /// <temperature_min>,
     /// <pressure>,
     /// <humidity>,
     /// <wind_speed>,
@@ -46,14 +43,35 @@ pub struct CliArgs {
     /// <sunrise>,
     /// <sunset>
     ///
-    /// Default: '<icon> <temperature_celsius>°C'
-    #[structopt(short, long)]
+    /// Default: '<icon> <temperature>°C'
+    #[structopt(short = "f", long)]
     pub format: Option<String>,
     /// The interval in seconds how often the weather status is updated
     ///
     /// If no interval is specified, wedder exits after printing the weather once
-    #[structopt(short, long)]
+    #[structopt(short = "i", long)]
     pub interval: Option<u64>,
+
+    /// The unit of temperature values
+    ///
+    /// Available units:
+    /// Celsius,
+    /// Fahrenheit,
+    /// Kelvin
+    ///
+    /// Default: Celsius
+    #[structopt(short = "t", long)]
+    pub temperature_unit: Option<Temperature>,
+    /// The unit for the wind speed
+    ///
+    /// Available units:
+    /// Ms,
+    /// Kmh,
+    /// Mph
+    ///
+    /// Default: Kmh
+    #[structopt(short = "s", long)]
+    pub wind_speed_unit: Option<WindSpeed>,
 
     /// The provider to use for pulling weather updates
     ///
@@ -62,7 +80,7 @@ pub struct CliArgs {
     /// OwmMock (for testing purposes)
     ///
     /// Default: OpenWeatherMap
-    #[structopt(short, long)]
+    #[structopt(short = "w", long)]
     pub weather_provider: Option<WeatherProvider>,
     /// The API key for the corresponding weather provider
     #[structopt(short = "k", long)]
@@ -75,11 +93,8 @@ pub struct CliArgs {
     /// Manual
     ///
     /// Default: Ip
-    #[structopt(short, long)]
+    #[structopt(short = "l", long)]
     pub location_provider: Option<LocationProvider>,
-    /// Prints the current city
-    #[structopt(short = "C", long)]
-    pub current_city: bool,
     /// Latitude of the location to display the weather status for
     #[structopt(long)]
     pub lat: Option<f32>,
@@ -93,14 +108,14 @@ impl CliArgs {
         if self.debug {
             self.debug()
         }
+        if self.current_city {
+            self.current_city()
+        }
         if self.default_config_path {
             if let Err(err) = self.default_config_path() {
                 println!("{}", err.to_string());
                 process::exit(1);
             }
-        }
-        if self.current_city {
-            self.current_city()
         }
     }
 
@@ -112,16 +127,6 @@ impl CliArgs {
         }
         debug!("Read {:?}", self);
     }
-
-    fn default_config_path(&self) -> Result<()> {
-        let path = config::file::default_config_path()
-            .ok_or("Couldn't get default config path")?
-            .to_str().map(std::string::ToString::to_string)
-            .ok_or("Couldn't parse default config path")?;
-        println!("{}", path);
-        process::exit(0);
-    }
-
     fn current_city(&self) {
         match IpApi::new().location() {
             Ok(location) => {
@@ -134,5 +139,14 @@ impl CliArgs {
                 process::exit(1)
             }
         }
+    }
+
+    fn default_config_path(&self) -> Result<()> {
+        let path = config::file::default_config_path()
+            .ok_or("Couldn't get default config path")?
+            .to_str().map(std::string::ToString::to_string)
+            .ok_or("Couldn't parse default config path")?;
+        println!("{}", path);
+        process::exit(0);
     }
 }